# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Localhost server"
        },
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local server"
        },
        {
            "url": "https://api.example.com"
        }
    ],
    "paths": {
...
        "/hotels": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get available hotels by country and dates",
                "operationId": "54cf891af66ecd8bfd0a6d26b1cc8fc8",
                "parameters": [
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Country to filter hotels (optional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Finlandia"
                        }
                    },
                    {
                        "name": "check_in",
                        "in": "query",
                        "description": "Check-in date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-06-01"
                        }
                    },
                    {
                        "name": "check_out",
                        "in": "query",
                        "description": "Check-out date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-06-10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Available hotels retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Available hotels": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Hotels retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hotels found in the selected location"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/hotels/{id}/rooms": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": "Get available rooms for a hotel by dates",
                "operationId": "fadd739494767aaedc96a7d925e5f245",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the hotel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "check_in",
                        "in": "query",
                        "description": "Check-in date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-06-01"
                        }
                    },
                    {
                        "name": "check_out",
                        "in": "query",
                        "description": "Check-out date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-06-10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Available rooms retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rooms retrieved successfully"
                                        },
                                        "available_rooms": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hotel not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format: Bearer {token}"
            }
        }
    },
...
}